+++ client/public/index.html
@@ -23,7 +23,7 @@
     <meta name="keywords" content="Voxo">
     <meta name="author" content="Voxo">
     <link rel="icon" href="assets/images/favicon/2.png" type="image/x-icon">
-    <title>404</title>
+    <title>Voho</title>
 
     <!--Google font-->
     <link rel="preconnect" href="https://fonts.gstatic.com">
