+++ client/src/App.js
@@ -1,8 +1,8 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Navbar from "./common/Navbar";
-import Footer from "./common/Footer";
-import PageNotFound from "./pages/404/PageNotFound";
+import Footer from "./common/footer/Footer";
+import Navbar from "./common/navbar/Navbar";
 import Home from "./pages/home/model";
+import PageNotFound from './pages/404/PageNotFound'
 
 function App() {
   return (
@@ -10,9 +10,9 @@ function App() {
       <Navbar />
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="*" element={<PageNotFound />} />
+        <Route path="*" element={<PageNotFound/>} />
       </Routes>
-      <Footer />
+      <Footer/>
     </BrowserRouter>
   );
 }
