--- client/src/_redirects
@@ -1,1 +1,0 @@
-/*  /index.html  200
